+++ doc/whatsnew.rst
@@ -14,6 +14,7 @@ Next release
 Bug fixes
 ---------
 
+- :obj:`NoSpecifiedRelationship` and :obj:`PrimaryMeasureRelationship` do not need to be instantiated; they are singletons (:issue:`54`).
 - `attributes=` "d" ignored in :func:`.to_pandas` (:issue:`55`).
 
 v2.0.0 (2021-01-26)